Rename pry-cyrano to pry-byetypo
