+++ lib/pry-byetypo/base.rb
