+++ lib/pry-byetypo/exceptions/active_record/base.rb
