+++ lib/pry-byetypo/exceptions/active_record/configuration_error.rb
