+++ lib/pry-byetypo/exceptions/active_record/statement_invalid.rb
