+++ lib/pry-byetypo/exceptions/exceptions_base.rb
