+++ lib/pry-byetypo/exceptions/name_error.rb
