+++ lib/pry-byetypo/exceptions_handler.rb
