+++ lib/pry-byetypo/setup/application_dictionary.rb
