+++ lib/pry-byetypo/setup/checks/base.rb
