+++ lib/pry-byetypo/setup/checks/database_pool.rb
@@ -4,18 +4,18 @@ require_relative "base"
 
 module Setup
   module Checks
-    class DatabaseUrl < Base
+    class DatabasePool < Base
       class << self
         def name
-          "url"
+          "pool"
         end
 
         def variable
-          "DATABASE_URL"
+          "DATABASE_POOL"
         end
 
         def missing_variable_msg
-          "[PRY-CYRANO] ENV[\"DATABASE_URL\"] is empty. Please assign a value to it to enable the functionality of pry-cyrano."
+          "[PRY-BYETYPO] ENV[\"DATABASE_POOL\"] is empty. Please assign a value to it to enable the functionality of pry-byetypo."
         end
       end
     end
