+++ lib/pry-byetypo/version.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class Pry
-  module Cyrano
+  module Byetypo
     VERSION = "1.0.0"
   end
 end
