+++ sig/pry/byetypo.rbs
@@ -1,5 +1,5 @@
 module Pry
-  module Cyrano
+  module Byetypo
     VERSION: String
     # See the writing guide of rbs: https://github.com/ruby/rbs#guides
   end
