+++ spec/pry/byetypo_spec.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RSpec.describe Pry::Byetypo do
+  it "has a version number" do
+    expect(Pry::Byetypo::VERSION).not_to be nil
+  end
+end
