--- spec/pry/cyrano_spec.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-RSpec.describe Pry::Cyrano do
-  it "has a version number" do
-    expect(Pry::Cyrano::VERSION).not_to be nil
-  end
-end
